Add Ignition Isolation
Found Ignition wire to be normally high, even when accessory power is not turned on. This was causing leakage current on the IgnitionPin input to partially power the uC when accessory power was off. Added a transistor to the circuit to use a digital switch to isolate that input until the uC is ready for it.